refactor rate a course button to its own component
